apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "FastAPI backend for AI Lead Qualification System"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "livekit-api>=0.7",
    "python-dotenv",
    "shared",
    "context-builder",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    # Billing
    "stripe>=7.0",
    # Email validation
    "email-validator>=2.0",
    # HTTP client for verification APIs
    "httpx>=0.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
